+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-355](https://github.com/itk-dev/hoeringsportal/pull/355)
+  Added citizen access check.
+
 ## [3.2.0] - 2023-08-23 - Citizen proposal
 
 * [PR-353](https://github.com/itk-dev/hoeringsportal/pull/353)